% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{my_dataset}
\alias{my_dataset}
\title{RameauEnv_Foret2UNIMARC2}
\format{
A data frame with 680 rows and 244 variables
}
\source{
GallicaEnv
}
\usage{
my_dataset
}
\description{
Jeu de donnees bibliographiques.
}
\keyword{datasets}
